Define assembly version attributes so that we can reference them from multiple projects
