--- source/Octopus.Tentacle/Versioning/AssemblyGitBranchAttribute.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace Octopus.Tentacle.Versioning
-{
-    [AttributeUsage(AttributeTargets.Assembly)]
-    public sealed class AssemblyGitBranchAttribute : Attribute
-    {
-        public AssemblyGitBranchAttribute(string branchName)
-        {
-            BranchName = branchName;
-        }
-
-        public string BranchName { get; }
-    }
-}
